fix has_manifest for 404 responses - do not evaluate_media_type()
If getting a 404 response when calling 'has_manifest' then if the content-type
returned for the json error body is 'application/json; charset=utf-8'
then the call to 'evaluate_media_type' will fail because it does not expect
charset=utf-8 to be valid here. see discussion in:
https://github.com/camallo/dkregistry-rs/pull/113

This change follows the approach taken in other methods above which is to
match on 'status' before continuing.